switch RollCount to int and improve numeric input logic

+++ src/backend/VoltStream.Application/Features/Supplies/Commands/UpdateSupplyCommand.cs
@@ -15,7 +15,7 @@ public record UpdateSupplyCommand(
     long ProductId,
     string ProductName,
     string Unit,
-    decimal RollCount,
+    int RollCount,
     decimal LengthPerRoll,
     decimal TotalLength,
     decimal UnitPrice,
