+++ src/backend/VoltStream.Application/Features/WarehouseStocks/DTOs/WarehouseStockDto.cs
@@ -5,7 +5,7 @@ using VoltStream.Application.Features.Products.DTOs;
 public record WarehouseStockDto
 {
     public long Id { get; set; }
-    public decimal RollCount { get; set; }  // rulon soni
+    public int RollCount { get; set; }  // rulon soni
     public decimal LengthPerRoll { get; set; }  //rulon uzunligi
     public decimal TotalLength { get; set; } // jami uzunlik
     public decimal UnitPrice { get; set; }
