+++ src/backend/VoltStream.Application/Features/WarehouseStocks/DTOs/WarehouseStockForProductDto.cs
@@ -3,7 +3,7 @@
 public record WarehouseStockForProductDto
 {
     public long Id { get; set; }
-    public decimal RollCount { get; set; }
+    public int RollCount { get; set; }
     public decimal LengthPerRoll { get; set; }
     public decimal TotalLength { get; set; }
     public decimal UnitPrice { get; set; }
