+++ src/frontend/VoltStream.WPF/Sales/ViewModels/WarehouseStockViewModel.cs
@@ -9,7 +9,7 @@ public partial class WarehouseStockViewModel : ViewModelBase
     public long Id { get; set; }
     public long WarehouseId { get; set; }
 
-    [ObservableProperty] private decimal rollCount;
+    [ObservableProperty] private int rollCount;
     [ObservableProperty] private decimal lengthPerRoll;
     [ObservableProperty] private decimal totalLength;
     [ObservableProperty] private decimal unitPrice;
